+++ aldrin/src/call.rs
@@ -4,18 +4,21 @@ use crate::handle::Handle;
 use crate::promise::Promise;
 use std::fmt;
 use std::task::{Context, Poll};
+use std::time::Instant;
 
 /// Pending call.
 pub struct Call<Args, T: ?Sized, E: ?Sized> {
     id: u32,
+    timestamp: Instant,
     args: Option<Args>,
     promise: Option<Promise<T, E>>,
 }
 
 impl<Args, T: ?Sized, E: ?Sized> Call<Args, T, E> {
-    pub(crate) fn new(id: u32, args: Args, promise: Promise<T, E>) -> Self {
+    pub(crate) fn new(id: u32, timestamp: Instant, args: Args, promise: Promise<T, E>) -> Self {
         Self {
             id,
+            timestamp,
             args: Some(args),
             promise: Some(promise),
         }
@@ -24,6 +27,11 @@ impl<Args, T: ?Sized, E: ?Sized> Call<Args, T, E> {
     /// Returns the call's function id.
     pub fn id(&self) -> u32 {
         self.id
+    }
+
+    /// Returns the timestamp when the call was received.
+    pub fn timestamp(&self) -> Instant {
+        self.timestamp
     }
 
     /// Returns a reference to the call's arguments.
@@ -99,6 +107,7 @@ impl<Args, T: ?Sized, E: ?Sized> Call<Args, T, E> {
     pub fn cast<T2: ?Sized, E2: ?Sized>(self) -> Call<Args, T2, E2> {
         Call {
             id: self.id,
+            timestamp: self.timestamp,
             args: self.args,
             promise: self.promise.map(Promise::cast),
         }
@@ -328,6 +337,7 @@ where
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         f.debug_struct("Call")
             .field("id", &self.id)
+            .field("timestamp", &self.timestamp)
             .field("args", &self.args)
             .field("promise", &self.promise)
             .finish()