+++ aldrin/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- The client now records the timestamp when a call is received. This is available via the new
+  functions `Call::timestamp()` and `low_level::Call::timestamp()`.
+
 ## [0.11.0] - 2025-01-07
 
 ### Added